requirements.txt
BeautifulSoup==3.2.1
Django==1.4.3
MySQL-python==1.2.3
PAM==0.4.2
PIL==1.1.7
Pillow==2.3.0
PyICU==1.4
South==0.7.6
Whoosh==0.3.18
anyjson==0.3.3
apt-xapian-index==0.44
aptoncd==0.1.98-bzr117-1.2
boto==2.13.3
celery==3.0.20
dj-database-url==0.2.0
django-celery==3.0.17
django-common-helpers==0.3
django-cron==0.2.2
django-evolution==0.6.9
django-extensions==1.1.1
django-filer==0.9.5
django-import-export==0.1.4
django-model-report==0.1.5
django-mptt==0.5.5
django-nose==1.1
django-reversion==1.8.0
django-simple-captcha==0.4.2
django-simple-history==1.3.0
django-storages==1.1.8
django-suit==0.2.5
django-themes==0.0.5
djappsettings==0.1.0
djtables==0.1.2
duplicity==0.6.19
easy-thumbnails==1.4
html5lib==0.95
httplib2==0.7.4
kombu==2.5.12
lxml==2.3.5
nose==1.2.1
oauth==1.0.1
paramiko==1.7.7.1
pexpect==2.4
pisa==3.0.33
piston-mini-client==0.7.2
protobuf==2.4.1
psycopg2==2.4.5
pyOpenSSL==0.13
pycrypto==2.6
pycups==1.9.62
pycurl==7.19.0
pygraphviz==1.1
pyinotify==0.9.3
pyserial==2.5
pysmbc==1.0.13
python-apt==0.8.7ubuntu4.1
python-dateutil==2.1
redis==2.7.6
reportlab==2.5
simplejson==2.6.1
six==1.6.1
sorl-thumbnail==11.12
system-service==0.1.6
tablib==0.9.11
uTidylib==0.2
vboxapi==1.0
virtualenv==1.7.1.2
wsgiref==0.1.2
xlrd==0.9.2
xlwt==0.7.4
